refactor: move tournament create to /pages folder and slightly adjust styling
