+++ src/app/modules/tournament-management/components/management-router/management-router.component.ts
@@ -28,7 +28,7 @@ export class ManagementRouterComponent implements OnInit {
 		{ icon: 'computer', header: 'local', link: 'local-tournaments' },
 		{ icon: 'language', header: 'published', link: 'published-tournaments', showIfObservable: this.isLoggedIn$ },
 		{ type: 'divider' },
-		{ icon: 'add', header: 'create', link: '/tournament-management/tournament-overview' },
+		{ icon: 'add', header: 'create', link: 'create' },
 		{ icon: 'list', header: 'all tournaments', link: '/tournament-management/tournament-overview', showIfObservable: this.isLoggedIn$ },
 		{ icon: 'admin_panel_settings', header: 'administrator', link: '/tournament-management/tournament-overview', showIfObservable: this.isTournamentManager$ },
 	];
