--- src/app/modules/tournament-management/components/tournament-manage/tournament-create/tournament-create.component.scss
@@ -1,10 +1,0 @@
-.extra-padding {
-	margin-bottom: 15px;
-}
-
-.scroll {
-	padding-right: 8px;
-
-	height: 100%;
-	overflow-y: auto;
-}
