+++ src/app/modules/tournament-management/pages/tournament-create/tournament-create.component.scss
@@ -1,0 +1,3 @@
+.tournament-create {
+	padding: 8px;
+}
