+++ src/app/modules/tournament-management/tournament-management-routing.module.ts
@@ -3,7 +3,7 @@ import { RouterModule, Routes } from '@angular/router';
 import { ManagementRouterComponent } from './components/management-router/management-router.component';
 import { TournamentAllPublishedAdministratorComponent } from './components/tournament-view/tournament-all-published-administrator/tournament-all-published-administrator.component';
 import { TournamentAllPublishedComponent } from './components/tournament-view/tournament-all-published/tournament-all-published.component';
-import { TournamentCreateComponent } from './components/tournament-manage/tournament-create/tournament-create.component';
+import { TournamentCreateComponent } from './pages/tournament-create/tournament-create.component';
 import { TournamentMyPublishedComponent } from './components/tournament-view/tournament-my-published/tournament-my-published.component';
 import { TournamentOverviewComponent } from './components/tournament-view/tournament-overview/tournament-overview.component';
 import { TournamentEditComponent } from './components/tournament-manage/tournament-edit/tournament-edit.component';
@@ -14,14 +14,14 @@ const routes: Routes = [
 	{
 		path: '', component: ManagementRouterComponent, children: [
 			{ path: 'tournament-overview', component: TournamentOverviewComponent },
-			{ path: 'tournament-overview/tournament-create', component: TournamentCreateComponent },
 			{ path: 'tournament-overview/tournament-edit/:id/:published', component: TournamentEditComponent },
 			{ path: 'tournament-overview/tournament-my-published', component: TournamentMyPublishedComponent },
 			{ path: 'tournament-overview/tournament-all-published', component: TournamentAllPublishedComponent },
 			{ path: 'tournament-overview/tournament-all-published-administrator', component: TournamentAllPublishedAdministratorComponent },
 
 			{ path: 'local-tournaments', component: LocalTournamentsComponent },
-			{ path: 'published-tournaments', component: PublishedTournamentsComponent }
+			{ path: 'published-tournaments', component: PublishedTournamentsComponent },
+			{ path: 'create', component: TournamentCreateComponent }
 		]
 	}
 ];
